including changes in model of working

+++ src/EProductivity.Core/Model/Organization.cs
@@ -5,7 +5,7 @@ namespace EProductivity.Core.Model
 {
     public class Organization
     {
-        public int OrganizationId { get; set; }
+        public long OrganizationId { get; set; }
         public string Document { get; set; }
         public OrganizationType Type { get; set; }
         public List<EProductivityUser> Users { get; set; }
